Add TemplateResult type to pass along creation results to caller
This change fixes PowerShell/vscode-powershell#390 which states that the
New Project experience in VS Code doesn't work well with project paths
which use characters like '~' to represent the home directory.  This
change causes the fully resolved DestinationPath of the template creation
to be passed back to the editor so that it can be opened.