Create AnalysisServiceProvider class
Our previous attempts to prevent crash caused by `Invoke-ScriptAnalyzer` outputting on a thread different than that of its invocation have failed. So, in this approach we create a class that analysis services and soon to be created formatting service can instantiate to avail PSScriptAnalyzer's services. This will keep only 1 runspace for each service and thus potentially prevent the crash.